split functions as per role
